// IBEXDATA/src/app/models/Contractors.model.ts
export class Contractors {
    contractorId!: number;
    contractorIdentity?: string;//mispar c.p
    contractorName?: string;///
    managementName?: string;///
    managementId?: number;///
    address?: string;///
    certificateConsortium?: string;//t.eged
    form50?: string;///
     insertDate?: Date;
     updateDate?: Date;
     contractorStatus!: number;
  }